refactor:#265 auth find-password 리팩토링
