+++ src/app/(auth)/find-password/_components/step-password.tsx
