+++ src/app/(auth)/find-password/_types/funnel.type.ts
